requirements.txt
hrml==1.0.2
numpy==1.19.2
jinja2==2.11.2
PyGithub==1.55
requests==2.24.0
Flask==1.1.2
Werkzeug==1.0.1
opencv-python==3.4.2.16